# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Auctra API
  version: "2026-07-01"
  summary: On-chain English auctions, over HTTP.
  description: |
    Auctra runs sealed-clock English auctions that settle on Ethereum. This API
    gives you the auction house without the blockchain plumbing: list an item,
    watch bids arrive over webhooks, settle, and pay out — in JSON.

    **The one rule that shapes everything.** A bid moves the bidder's own ETH,
    so only the bidder's private key can authorise it. Auctra never holds that
    key. Write endpoints return an unsigned `transaction_request`; you sign it
    with the user's wallet and broadcast it, either yourself or through
    `POST /v1/transactions`.

    Amounts are decimal strings of wei. Never JSON numbers — 1 ETH is 10^18 wei,
    and a double loses integer precision above 2^53.
  termsOfService: https://auctra.dev/terms
  contact:
    name: Auctra developer support
    url: https://docs.auctra.dev
    email: developers@auctra.dev
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://sandbox-api.auctra.dev
    description: Sandbox — Sepolia testnet. Test ETH, real code paths.
  - url: https://api.auctra.dev
    description: Production — Ethereum mainnet.
security:
  - bearerAuth: []
...
paths:
...
  /v1/refunds:
    get:
      tags: [Refunds]
      summary: List refund credits
      description: |
        When a bidder is outbid, the contract credits them rather than sending
        ETH back. Pushing a refund would let a malicious contract revert on
        receive and freeze the auction for everyone. Refunds are pulled.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - name: bidder
          in: query
          schema: { $ref: "#/components/schemas/Address" }
        - name: withdrawn
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: A paginated list of refund credits.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/List"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Refund" }
  /v1/refunds/balance:
    get:
      tags: [Refunds]
      summary: Read a withdrawable balance
      description: Reads `pendingReturns()` straight from the chain, not the cache.
      parameters:
        - name: bidder
          in: query
          required: true
          schema: { $ref: "#/components/schemas/Address" }
      responses:
        "200":
          description: The live withdrawable balance.
          content:
            application/json:
              example:
                object: refund_balance
                bidder: "0x9c8b7a6f5e4d3c2b1a0f9e8d7c6b5a4f3e2d1c0b"
                withdrawable_wei: "1050000000000000000"
                withdrawable_display: "1.05"
                source: chain
  /v1/refunds/withdraw:
    post:
      tags: [Refunds]
      summary: Withdraw a refund balance
      description: |
        Prepares `withdraw()`, which sweeps the bidder's entire credited
        balance across all auctions in a single transaction.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bidder]
              properties:
                bidder: { $ref: "#/components/schemas/Address" }
      responses:
        "200":
          description: Withdrawal transaction prepared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { type: string, const: withdrawal_intent }
                  bidder: { $ref: "#/components/schemas/Address" }
                  transaction_request: { $ref: "#/components/schemas/TransactionRequest" }
  /v1/transactions:
    post:
      tags: [Transactions]
      summary: Broadcast a signed transaction
      description: |
        Optional. You can send a signed transaction to any node; this endpoint
        exists so an integration needs one network dependency instead of two.

        Returns `202 Accepted`, not `200`. The chain decides whether a
        transaction succeeds, and it has not decided yet.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signed_transaction]
              properties:
                signed_transaction:
                  type: string
                  pattern: "^0x[a-fA-F0-9]+$"
      responses:
        "202":
          description: Accepted into the mempool.
          content:
            application/json:
              example:
                object: transaction
                hash: "0x4c1d7a90f3b2e5c8a1d4f7b0e3c6a9d2f5b8e1c4a7d0f3b6e9c2a5d8f1b4e7c0"
                status: pending
                confirmations: 0
                poll: /v1/transactions/0x4c1d...
        "422": { $ref: "#/components/responses/ChainError" }
  /v1/transactions/{hash}:
    get:
      tags: [Transactions]
      summary: Retrieve a transaction
      parameters:
        - name: hash
          in: path
          required: true
          schema: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
      responses:
        "200":
          description: |
            `status` is `pending` until mined, then `confirmed` or `reverted`.
            A `reverted` transaction still consumed gas.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
  /v1/webhook_endpoints:
    get:
      tags: [Webhooks]
      summary: List webhook endpoints
      responses:
        "200":
          description: Your endpoints. Secrets are never returned after creation.
          content:
            application/json:
              example:
                object: list
                data:
                  - object: webhook_endpoint
                    id: whe_6b3d0f8a2c1e
                    url: https://example.com/hooks/auctra
                    enabled_events: [auction.bid_placed, auction.outbid, auction.settled]
                    status: enabled
                    created_at: "2026-07-01T08:14:22.417Z"
                has_more: false
                next_cursor: null
    post:
      tags: [Webhooks]
      summary: Create a webhook endpoint
      description: |
        The response contains `secret` exactly once. Store it; every subsequent
        request is signed with it, and you must verify that signature before
        trusting the body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, enabled_events]
              properties:
                url: { type: string, format: uri }
                enabled_events:
                  type: array
                  items: { $ref: "#/components/schemas/EventType" }
            example:
              url: https://example.com/hooks/auctra
              enabled_events: [auction.bid_placed, auction.outbid, auction.settled]
      responses:
        "201":
          description: Endpoint created. `secret` is shown once.
          content:
            application/json:
              example:
                object: webhook_endpoint
                id: whe_6b3d0f8a2c1e
                url: https://example.com/hooks/auctra
                enabled_events: [auction.bid_placed, auction.outbid, auction.settled]
                secret: whsec_9d4f2a7c1b8e0356af914d2c7b0e5836a1f4c9d2
                created_at: "2026-07-01T08:14:22.417Z"
  /v1/webhook_endpoints/{id}:
    delete:
      tags: [Webhooks]
      summary: Delete a webhook endpoint
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted.
  /v1/events:
    get:
      tags: [Webhooks]
      summary: List events
      description: |
        Every webhook we send is also a durable row here. If your listener was
        down, don't ask us to redeliver — read the log and reconcile.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/EventType" }
      responses:
        "200":
          description: A paginated list of events.
          content:
            application/json:
              example:
                object: list
                data:
                  - object: event
                    id: evt_7c2e9a1f4b6d
                    type: auction.outbid
                    created_at: "2026-07-09T10:31:04.882Z"
                    data:
                      object:
                        auction: auc_3f9a2c7b1d4e
                        outbid_bidder: "0x9c8b7a6f5e4d3c2b1a0f9e8d7c6b5a4f3e2d1c0b"
                        refund_wei: "1050000000000000000"
                has_more: true
                next_cursor: evt_7c2e9a1f4b6d
  /v1/events/{id}:
    get:
      tags: [Webhooks]
      summary: Retrieve an event
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The event.
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer sk_test_...` in sandbox,
        `Authorization: Bearer sk_live_...` in production.

        A live key sent to the sandbox host fails with `environment_mismatch`,
        not `invalid_api_key` — mixing environments is the most common
        integration bug and the error should say so.

        Keys are stored as salted hashes. We cannot recover a lost key; rotate it.
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Page size, 1–100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    StartingAfter:
      name: starting_after
      in: query
      description: |
        An object id to page after. Cursor, not offset: auctions are inserted
        constantly, and `?page=2` would skip or duplicate rows whenever a new
        one lands between requests.
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        A unique string, e.g. a UUID. Replaying the same key with the same body
        returns the original response verbatim; with a different body it
        returns `409 idempotency_key_reused`. Keys expire after 24 hours.
      schema: { type: string, maxLength: 255 }
  schemas:
    Address:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
      description: A 20-byte Ethereum address, lowercased in responses.
      examples: ["0x7a3f1c8b9d2e4f6a0b1c2d3e4f5a6b7c8d9e0f1a"]
    Wei:
      type: string
      pattern: "^[0-9]+$"
      description: |
        A decimal string of wei. Always a string: 1 ETH = 10^18 wei, and JSON
        numbers are doubles, which lose integer precision above 2^53.
      examples: ["1050000000000000000"]
    EventType:
      type: string
      enum:
        - auction.created
        - auction.bid_placed
        - auction.outbid
        - auction.extended
        - auction.ending_soon
        - auction.ended
        - auction.settled
        - auction.cancelled
        - refund.credited
        - refund.withdrawn
        - item.claimed
    List:
      type: object
      properties:
        object: { type: string, const: list }
        has_more: { type: boolean }
        next_cursor:
          type: [string, "null"]
          description: Pass as `starting_after` to fetch the next page. `null` on the last page.
    TransactionRequest:
      type: object
      description: |
        An unsigned transaction. Sign it with the user's wallet — Auctra never
        sees a private key. `gas_limit` already includes 20% headroom.
      properties:
        object: { type: string, const: transaction_request }
        chain_id: { type: integer }
        to: { $ref: "#/components/schemas/Address" }
        data: { type: string }
        value: { $ref: "#/components/schemas/Wei" }
        gas_limit: { type: [string, "null"] }
        max_fee_per_gas: { type: [string, "null"] }
        max_priority_fee_per_gas: { type: [string, "null"] }
        nonce_hint: { type: [integer, "null"] }
...
    Refund:
      type: object
      properties:
        object: { type: string, const: refund }
        id: { type: string }
        bidder: { $ref: "#/components/schemas/Address" }
        amount_wei: { $ref: "#/components/schemas/Wei" }
        amount_display: { type: string }
        withdrawn: { type: boolean }
        withdrawal_transaction: { type: [string, "null"] }
        chain_id: { type: integer }
        created_at: { type: string, format: date-time }
      examples:
        - object: refund
          id: ref_1e5a8c3f7b2d
          bidder: "0x9c8b7a6f5e4d3c2b1a0f9e8d7c6b5a4f3e2d1c0b"
          amount_wei: "1050000000000000000"
          amount_display: "1.05"
          withdrawn: false
          withdrawal_transaction: null
          chain_id: 11155111
          created_at: "2026-07-08T14:31:52.006Z"
    Transaction:
      type: object
      properties:
        object: { type: string, const: transaction }
        hash: { type: string }
        status: { type: string, enum: [pending, confirmed, reverted] }
        block_number: { type: [integer, "null"] }
        confirmations: { type: integer }
        gas_used: { type: [string, "null"] }
        effective_gas_price: { type: [string, "null"] }
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [type, code, message, request_id]
          properties:
            type:
              type: string
              enum:
                - authentication_error
                - permission_error
                - invalid_request_error
                - chain_error
                - idempotency_error
                - rate_limit_error
                - api_error
            code:
              type: string
              description: Switch on this. Messages change; codes don't.
            message: { type: string }
            param: { type: string }
            detail: { type: object }
            request_id:
              type: string
              description: Quote this in a support request. It's the only string we need.
            docs_url: { type: string, format: uri }
  responses:
...
    NotFound:
      description: No such object under this API key's organisation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ChainError:
      description: |
        The chain rejected the transaction. `message` carries the decoded custom
        error from AuctionHouse.sol, not "execution reverted".
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...